requirements.txt
azure-functions>=1.18.0
azure-storage-blob>=12.0.0
azure-identity>=1.15.0
azure-monitor-ingestion>=1.0.0
azure-core>=1.29.0
requests>=2.28.0
